requirements.txt
absl-py==1.4.0
alembic==1.11.1
astor==0.8.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
contextlib2==21.6.0
cucim==23.6.0
cycler==0.11.0
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
docker==6.1.3
einops==0.6.1
entrypoints==0.4
fastrlock==0.8.1
filelock==3.11.0
fire==0.5.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.8.0
huggingface-hub==0.16.4
humanize==4.6.0
idna==3.4
ignite==1.1.0
imagecodecs==2021.11.20
imageio==2.31.1
importlib-metadata==5.2.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json-tricks==3.17.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.0
lmdb==1.4.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==1.30.1
monai==1.1.0
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==4.0.2
ninja==1.11.1
nni==2.10.1
nptyping==2.5.0
numexpr==2.8.4
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.8.0.74
openslide-python==1.1.2
opt-einsum==3.3.0
optuna==3.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prettytable==3.7.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydicom==2.4.2
Pygments==2.15.1
PyJWT==2.8.0
pynrrd==1.0.0
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.10
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==25.1.0
querystring-parser==1.2.4
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.23.1
rich==13.4.2
rsa==4.9
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
shellingham==1.5.0.post1
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
tables==3.7.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.12.1
torch==1.13.1
torchio==0.18.92
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.21.3
typeguard==2.13.3
typer==0.9.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
